baseline_optim/requirements.txt
absl-py==1.0.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
Deprecated==1.2.13
docker-pycreds==0.4.0
filelock==3.5.1
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
imageio==2.14.1
importlib-metadata==4.10.1
importlib-resources==5.4.0
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.3.2
llvmlite==0.38.0
Markdown==3.3.6
matplotlib==3.5.1
msgpack==1.0.3
networkx==2.6.3
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.0
pathtools==0.1.2
Pillow==9.0.0
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynndescent==0.5.6
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
ray==1.10.0
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.5.6
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
split-folders==0.4.3
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2022.2.2
torch==1.10.2
torch-tb-profiler==0.3.1
torchvision==0.11.3
tqdm==4.62.3
typing_extensions==4.0.1
umap-learn==0.5.2
urllib3==1.26.8
wandb==0.12.10
Werkzeug==2.0.2
wrapt==1.13.3
yaspin==2.1.0
zipp==3.7.0
